Use int type hints on IntegerAttribute

+++ datafinder/src/datafinder/typed_attributes.py
@@ -33,11 +33,11 @@ class IntegerAttribute(Attribute):
     def __init__(self, name: str, column_db_type: str, owner:str, parent=None):
         super().__init__(name, column_db_type, owner, parent)
 
-    def eq(self, value: float) -> Operation:
+    def eq(self, value: int) -> Operation:
         return PrimitiveEqOperation(self, value)
 
-    def __eq__(self, value: float) -> Operation:
+    def __eq__(self, value: int) -> Operation:
         return PrimitiveEqOperation(self, value)
 
-    def __gt__(self, value: float) -> Operation:
+    def __gt__(self, value: int) -> Operation:
         return PrimitiveGreaterThanOperation(self, value)
